Updated instructions and Exercises0
I made instructions more clear with an example of the tree I want to obtain. I also created the file Exercise0_BELLINO.ipynb where I can do the exercise and then load it on the repository of the course. It is precisely the only file that I have to include in the branch STAMET_FS25 before committing and pushing.