Instructions changed
After having seen what a mess the STAMET_FS25 branch is, I don't want to merge it with Personal anymore. I'd rather follow the same procedure that I intended to adopt to selectively add the solved exercise sheets into the common branch to pick only the new Exercises and the Solutions from the common branch.