Possible end of Exercise 4.1
The results are there, they make sense. Unfortunately the graph of the likelihood in the case of correlated luminosity errors is a bit weird: the steepness changes suddenly at the values corresponding to the two measurements. It kind of make sense, cause it raises faster between them, but still the discontinuity of the first derivative sounds suspicious.