Proceedings of 6.2
The first point of that exercise is now modified to perform the ANOVA test with a dedicated function. The same function is than used in the second part of the same exercise to compare pairs of samples. I am not sure that I can use the ANOVA for only two samples so now I will proceed with the Student-T test to see if they are compatible