Attempted background fit to a generic polynomial of order 3
Before hearing from Dr. Manzoni how to fit to a Chebyshev polynomial of order 3, I tried fitting the background to a generic polynomial of order 3 normalized between 2 and 6.
It is a complete failure: first, I found no smart way of constraining the polynomial to have only positive values (the absolute value makes the normalization impossible); second, maybe the peak in the data makes it difficult to fit the background only.
Gabri, if you want to work on it, please feel free to delete everything as it doesn't make much sense as of now.