Improved funtion description and point-wise crystall_ball
Now the function crystall_ball takes as an input a single point in [2,6] instead of a full array of points in that intervall. The drawback is that it is less flexible (it is now specifically normalized in [2,6] by construction) but it has the advantage of being more similar to the other functions.
I also added some descriptions to other functions and a brief explanation of what a crystall ball is and why we want to use it.