Solved crystal_ball normalization
First, the crystal_ball is now properly normalized (there was an issue regarding the point of switch between the gaussian and the polynomial tail, it was not -alpha but mu-alpha*sigma and thus we were integrating in the wrong intervals).
Second, crystal_ball is now able to handle both np.arrays and float as inputs.
Third, both the cheb3r and the crystal_ball have been integrated numerically in the intervall [2,6] and they are correctly normalized.