Model completed
As of now, the model describing the invariant mass spectrum is completed: we have developed the pdfs of the Chebyshev polynomials and the crystal ball; we have developed a function to unpack and manage the parameters; we have developed a function that sums the three features of the spectrum (background, JPsi resonance and Psi2S resonance) multipling each of them by the respective yields.
This function is the stepping stone for the next layer: a function that integrates this model inside the bins the data are divided into, providing the number of expected events inside each bin. This number can then be compared with the observed number to evaluate a Likelihood. We will probably assume Poissoin distribution of the number inside each bin.