Added integrated_model and clean up
The model part of the notebook should be completed: we now have a binned_model function that integrates the combined_model inside each bin when provided the bins' edges. Thus, it provides the number of expected events inside each bin, which can be compared directly to the observed number of events.
Also, this number can be used to compute the likelihood of the data with respect to the model. Then, the binned Negative Log Likelihood can be minimized to find the parameters that best describe the data. I am still in doubt regarding the NLL: shall we use the bNLL or the bENLL to factor in also the information coming from the total number of observed events? It can be compared simply to the sum of the yields of the bkg, the JPsi and the Psi2S resonances.

I have also removed the unnecessary cells to keep the notebook as tidy as possible.