Just a parameters issue
The model had negative values because of a bad choice of initial parameters for the Chebyshev polynomials. Apparently, the suggested ones in the README don't match our definition of Chebyshev polynomials combination so I used those that have been previously estimated by Gabri in a fit (check previous commits).
We still need to find a way to constrain the chebyshev polynomials to positive values... if they hit, even once, negative values, the bNLL can crash because of the logarithm.