End of dataclass transition
All the functions of pdfs and the combined model are now dataclasses, which unpacks the parameters, compute useful factors and evaluates the normalization factors only once, at the dataclass definition, and use them efficiently at each call of the function. In this way, the integration of the combined model insied each bin (which requires many many evaluation of combined_model) should be much faster (resulting in a speed up of the fit itself).
I have compared the speed of the fitting algorithm with same initial parameters until it stops because it runs into log of negative values: obviously, both the versions yield the same bNLLs (I checked the last one before nan) but while the function-based notebook runs it in 133s, the dataclass-based notebook only takes 1s.