Added combined model, transition to dataclasses
The existing pdfs are now combined by multiplying each one for the corresponding yield to produce the combined_model.
I have noticed that the function-based notebook takes a lot of time to evaluate the bNLL, mainly because the normalization of the crystal ball is computed again everytime it is called by the integrator quad (although the parameters are the same). I moved everything from functions to dataclasses, initialized with the parameters, having a callable function which evaluates the pdf or the combined model at the given point. This ensures that the normalization is computed only once, parameters are unpacked only once, and relevant factors like A,B of the poly_tail are calculated only once. This effort resulted in a great speed up: the fit bumps into problems after 133s for functions, after 1s for dataclasses (providing the same results before finding issues).
The fit is still useless because of Chebyshev polynomials: there are combinations of parameters that yield a negative pdf. We really have to discuss how to solve this issue.