Using a gaussian for the Psi(2s) resonance
The polynomial tail on the Psi(2s) resonance looked useless: not only it gave optimal parameters to the limits of the boundary, but they also have incredibly high standard deviations, affecting as a consequence the uncertainty on the cross section as well. For this reason, we use a simple gaussian to fit the resonance.