Run test completed:
Corrected a little misprint in in the probability function of the number
of runs, corrected the limits of the probability function, defined a
rejection interval to test the hypothesis that the data are randomly
distributed around the model (90% significance level). Plot and a
description of all operations done have been added at the beginning of
the Run Test section.