Starting task 4 (High statistics data still)
We have started working on the fourth and last task. We already manage to plot the graph showing the p-value observed and expected for different values of the cross-section. The package to plot the file is provided. Note that, as of now, only the High Statistic dataset is being used. In future, we will move to the LS dataset.