Fit with low stat data and upper limits:
A small description of passages in upper limit section has been added. Re-fit using low stat data (just repeated task one with all data in the window [2,6]). Added some code to compute the discovery significance and and upper limit to the cross section of Psi2s in the case of low stat.