Modified model for JPsi resonance
The JPsi resonance is now modeled as a crystal-ball plus a gaussian, both constrained to have the same mean but allowed to have different standard deviations and yields. Curiously, the yields come out very similarly, suggesting the half of the points are "low resolution" (scattering events resulting in mu mu emitted close to the propagation axis) and thus follow the distribution with the largest std, half are "high resolution" (mu mu produced almost perpendicularly to the axis of propagation) and follow the distribution with smallest std. It's worth noticing that the two optimized std are indeed different: one is 0.02, the other 0.05; when using a single crystal-ball, it was constrained to the middle value 0.03