Some changes suggested by Riccardo applied
Chi-squared back to ndf = nbins-parameters estimated. High statistics fit: restricted the window in 4.5 and chose a higher number of bins: nbins = 140. In this way the percentages of residuals in 1 and 2 sigmas are closer to 68 and 95 (70 and 96.4) BUT the shape of the distribution is stranger and a bit asymmetric. nbins = 150 solves this problem but the percentages get a bit worse. Verified that for such numbers of bins the minimum statistic per bin is large enough: for nbins = 150 is 38 (added one line of code to compute this). Should be enough. Modified the code for the two profile likelyhoods because they had not been updated to consider the gaussian for the psi2s and the crystall ball + gaussian for jpsi.