Improvements on the new binning method
I have found that setting 180 points per bin is produces pictures that are very nice, any number of points nevertheless yields very good results, all showing compatibility within the 95% CL. I adjusted the graphs so that we have close-ups of the JPsi resonance, both on the fit itself and the residuals. The normalized residuals appear to be really distributed as a gaussian.