Shitty commit to plot nice pictures
I have completely messed up the notebook, which is now in a terrible state. However, I managed to produce the nice plots we need for the presentation. If debugging on the analysis is needed, I strongly encourage you to go one commit back and start from there, instead of fixing this notebook.