Manually extracting the upper bounds
I am printing out the raw values of the curves plotted in the upper-limit graph and manually select the coordinates of the two points on different sides of the crossing point. Then, using a linear interpolation of them, I determine the value of the cross_section at the crossing point.