Updating both notebooks to the latest version
I have copied the notebooks from both branches Development and Alternative_JPsi. Now the notebooks here contain the latest version of our project, both using the simple model for the JPsi or the more complicated one with the sum of two distributions. After this commit we should be ready to merge our repository with the public one of the course.